Thermal instabilities analysis into the ALBA’s water storage tank and detailed description of the one-dimensional transient simulation algorithm
The water-cooling system of the ALBA accelerator has experienced thermal instability problems over the last years, which affect the control of the entrance temperature in its four main rings: Service Area, Booster, Storage and Experimental Hall. Initial studies carried out by an infrastructure work team have determined that the main source of this problem is an anomaly in the cold-water supply due to an external Co-generation Plant (ST4), on which ALBA depends to thermally regulate the Accelerator and the Air Conditioning systems.

In this article, the effect of the thermal instabilities in the hydraulic refrigeration circuit of the ALBA Synchrotron is studied at the D02 water tank. The analysis is performed by the use of a one-dimensional node-centered numerical simulation, carried out employing a self-made matlab code based on actual historical data. 

The heat transfer through the walls of the tank is quantified taking into account the thermal instabilities, and compared to the heat exchange behaviour in a steady state. The main objective of the paper is to study the necessity of insulating the tank, by the design of an adequate insulator which would reduce the heat loses in the tank. A transient simulation algorithm is presented and performed in order to solve the conduction heat transfer in the walls of the tank, solving the convection at their inner and outer sides.